requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
pillow
numpy
gunicorn
